Moving Success / Application Modals to sep folders
